pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redeploy"
version = "0.2.51"
description = "Infrastructure migration toolkit: detect → plan → apply"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "markdown-it-py>=3.0",
    "click>=8.0",
    "loguru>=0.7",
    "paramiko>=3.0",
    "httpx>=0.25",
    "rich>=13.0",
    "jmespath>=1.0",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.scripts]
redeploy = "redeploy.cli:cli"
redeploy-mcp = "redeploy.mcp_server:serve"

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]
op3 = ["op3>=0.1.8"]
mcp = ["mcp>=1.0"]

[tool.ruff]
line-length = 100

[tool.setuptools.packages.find]
where = ["."]
include = ["redeploy*"]

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
